Replace Widgets with library components and some fixes
