+++ apps/builddao/widget/Aside.jsx
@@ -1,3 +1,5 @@
+const { Button } = VM.require("buildhub.near/widget/components.button");
+
 const Container = styled.div`
   border-radius: 16px;
   border: 1px solid var(--Stroke-color, rgba(255, 255, 255, 0.2));
@@ -18,67 +20,20 @@ const Container = styled.div`
   }
 `;
 
-const TabButton = styled.button`
-  all: unset;
-  display: flex;
-  flex-shrink: 0;
-  padding: 8px 12px;
-  align-items: center;
-  gap: 10px;
-  align-self: stretch;
-
-  @media screen and (max-width: 768px) {
-    align-self: auto;
-  }
-
-  border-radius: 8px;
-  border: 1px solid var(--Stroke-color, rgba(255, 255, 255, 0.2));
-
-  color: var(--white-50, rgba(255, 255, 255, 0.7));
-
-  /* Body/14px */
-  font-size: 14px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: 170%; /* 23.8px */
-
-  transition: all 300ms;
-
-  &:hover,
-  &:active,
-  &.active {
-    border-radius: 8px;
-    background: var(--Yellow, #ffaf51) !important;
-
-    color: var(--black-100, #000) !important;
-    cursor: pointer;
-
-    /* Body/14px */
-    font-size: 14px;
-    font-style: normal;
-    font-weight: 400;
-    line-height: 170%; /* 23.8px */
-
-    svg {
-      filter: invert(1);
-    }
-  }
-`;
-
-const Aside = (props) => {
-  return (
-    <Container>
-      {Object.keys(props.routes || {}).map((route) => (
-        <TabButton
-          className={props.active === route && "active"}
-          onClick={() => props.setActiveRoute(route)}
-        >
-          <i className={`bi ${props.routes[route].icon}`}></i>
-          {props.routes[route].label}
-        </TabButton>
-      ))}
-    </Container>
-  );
-};
-
-return { Aside };
+return (
+  <Container>
+    {Object.keys(props.feeds || {}).map((feed) => (
+      <Button
+        variant={props.currentFeed === feed ? "primary" : "outline"}
+        onClick={() => props.setCurrentFeed(feed)}
+        className={
+          "align-self-stretch flex-shrink-0 justify-content-start fw-medium"
+        }
+        style={{ fontSize: "14px" }}
+      >
+        <i className={`bi ${props.feedsDict[feed].icon}`}></i>
+        {props.feedsDict[feed].label}
+      </Button>
+    ))}
+  </Container>
+);
