+++ apps/builddao/widget/resources-aside.jsx
@@ -1,3 +1,5 @@
+const { Button } = VM.require("buildhub.near/widget/components.button");
+
 const Container = styled.div`
   border-radius: 16px;
   border: 1px solid var(--Stroke-color, rgba(255, 255, 255, 0.2));
@@ -18,66 +20,21 @@ const Container = styled.div`
   }
 `;
 
-const TabButton = styled.button`
-  all: unset;
-  display: flex;
-  flex-shrink: 0;
-  padding: 8px 12px;
-  align-items: center;
-  gap: 10px;
-  align-self: stretch;
-
-  @media screen and (max-width: 768px) {
-    align-self: auto;
-  }
-
-  border-radius: 8px;
-  border: 1px solid var(--Stroke-color, rgba(255, 255, 255, 0.2));
-
-  color: var(--white-50, rgba(255, 255, 255, 0.7));
-
-  /* Body/14px */
-  font-size: 14px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: 170%; /* 23.8px */
-
-  transition: all 300ms;
-
-  &:hover,
-  &:active,
-  &.active {
-    border-radius: 8px;
-    background: var(--Yellow, #ffaf51) !important;
-
-    color: var(--black-100, #000) !important;
-    cursor: pointer;
-
-    /* Body/14px */
-    font-size: 14px;
-    font-style: normal;
-    font-weight: 400;
-    line-height: 170%; /* 23.8px */
-
-    svg {
-      filter: invert(1);
-    }
-  }
-`;
-
-console.log(props.currentResource);
-
 return (
   <Container>
     {props.resources.map((resource) => (
-      <TabButton
-        className={
-          props.currentResource === resource.name ? "active" : undefined
+      <Button
+        variant={
+          props.currentResource === resource.name ? "primary" : "outline"
         }
         onClick={() => props.setCurrentResource(resource)}
+        className={
+          "align-self-stretch flex-shrink-0 justify-content-start fw-medium"
+        }
+        style={{ fontSize: "14px" }}
       >
         <i className={`bi ${resource.biIcon}`}></i> {resource.name}
-      </TabButton>
+      </Button>
     ))}
   </Container>
 );
